Reduce instance store size to 8GB
In every AMI we have built so far except for the docker AMI, 8GB was
sufficiently large.  It may be possible to reduce the size further,
but for now I wanted to reduce it to 8GB.  This will make the AMIs
smaller and reduce the time required to wait for them to finish
encrypting.