Make TravisCI script correctly handle a version update
The way the code was written, if TERRAFORM_VERSION or PACKER_VERSION
was changed, the new version would not be updated unless it was newer
than the terraform.zip or packer.zip file saved in the curl cache
directory.  This may not always be the case, so I changed the code to
instead save the zipfile to the curl cache directory under its full
name.  This resolves this potential issue, since the full name
includes the version.  Therefore the curl command with the --time-cond
flag will always be comparing apples to apples.