Make sure that efs-utils is the last thing installed by Ansible
This role forces the installation of python2, which breaks any pip
modules that are installed afterwards because Ansible will see that
/usr/bin/python exists and it.  In particular, Ansible will try to
install python-apt during idempotence; since this package does not
exist in Debian 11 (Bullseye) that causes Ansible to fail.

I have created aws/efs-utils#57 to encourage the good folks at
aws/efs-utils to support python3 in their packaging.  (The code itself
already supports python3.)  Until they change that, we should be OK as
long as this role is the last thing that Ansible installs.