Remove .terraform caching and add explanatory comment
If we are caching the .terraform directory then we need to use the -upgrade
option to ensure the modules and plugins we use for linting are up-to-date.
However, the -upgrade=true option ignores what is already downloaded, and just
downloads the latest allowable versions. This defeats the purpose of caching,
so we will not bother.