Force a /usr/bin/python3 symlink to the install python
When Ansible runs it will look for /usr/bin/python3 and use it if it
finds it.  Therefore we need /usr/bin/python3 to point to the desired
python when packer runs.

This change will hose other tasks that expect to use the system
python at /usr/bin/python3, though, so we undo the change after
running packer.