Change the device name for the AMI and launch block mappings
I noticed that the root volume is the original 8GB volume from the
base AMI when packer attempts to build the AMI.  Looking in the AWS
console I saw that the EC2 instance where the AMI is being built has
an 8GB volume and a 20 GB volume attached at /dev/sda and /dev/xvda,
respectively.  Oddly, the attachment IDs are identical.  I believe
this is related to the fact that /dev/sda is the legacy name for
/dev/xvda, and somehow the /dev/sda mount is superseding the /dev/xvda
mount.