Add a test AMI build to the build workflow
The test build goes through the process of building an AMI but does
not actually create it.  This is an good thing to add since it will
catch changes that cause the AMI build process to fail.  Currently
these changes are only caught when we create a pre-release or release.