Change is_prerelease variable type
Switch from a string to a bool variable type to better reflect what the
variable represents. This also requires updating the build.yml GitHub
Actions workflow to remove variable value passing in the call to
Packer. Since this workflow does not create an AMI we do not need to
worry about the tag values since they will not be applied. The
github.event.release object also contains no useful values in that
workflow's context. The correct bool-like value for is_prerelease
should be passed in the prerelease.yml and release.yml workflows.

Co-authored-by: dav3r <david.redmin@trio.dhs.gov>
Co-authored-by: Shane Frasier <jeremy.frasier@trio.dhs.gov>