Force the Ansible provisioner to use sftp instead of scp
The latest Kali base AMI has an OpenSSH version greater than or equal
to 9.0 installed.  scp fails in that case because of a breaking
change, which is described here:
https://www.openssh.com/txt/release-9.0

This will be a problem for our other AMIs as other base AMIs upgrade
to newer versions of OpenSSH 9.

The use of sftp is the default behavior for Ansible, but Packer's
Ansible provisioner forces the use of scp by default as of this
commit:
hashicorp/packer@f760ab2