Set use_proxy to false for Ansible provisioners
This fixes our cisagov/*-packer builds, which have all been failing
recently now that Ubuntu 22.04 is the default for GitHub Actions'
runners.  See, for example, this failing build:
https://github.com/cisagov/samba-packer/actions/runs/3534584325
and this successful build:
https://github.com/cisagov/samba-packer/actions/runs/3541656743

The proxy uses an RSA key, which the newer version of OpenSSL in
GitHub Actions' Ubuntu 22.04 runner host now rejects as insecure.
Note that this same change was already made in cisagov/cyhy-amis#441.

It is also worth noting that use_proxy = false is the default setting
for Ansible; Packer explicitly overrides this default.