Update the process to set the neo4j password
In neo4j 5.0 the syntax for setting the password using the
set-initial-password command has changed. This commit changes the
password setting task into a block that will run whichever command is
appropriate based on the version of the neo4j package that is
installed on the system.

This change also adds a no_log: true parameter to both of the tasks to
set the password so that the password is not accidentally leaked if
there is an error.

This code is cribbed from the following commit by @mcdonnnj:
https://github.com/cisagov/ansible-role-bloodhound/pull/43/commits/ded4cb4f90441356c505afd2642681906e361138

Co-authored-by: Nick <50747025+mcdonnnj@users.noreply.github.com>