Remove unnecessary ami_block_device_mappings block
One only needs to specify ami_block_mappings_block to add additional
volumes when the AMI launches that are not present when the instance
on which the AMI is built is launched:
https://developer.hashicorp.com/packer/plugins/builders/amazon/ebs#ami-block-device-mappings-example

I suppose another use might be if you want the root disk when the AMI
is launched to be different than it is when the instance on which the
AMI is being built is launched, but that doesn't apply here either.

It is certainly the case that there is no need to configure the same
root volume in the same way in both the launch_block_device_mappings
and ami_block_device_mappings blocks.