Sort the keys in the Dependabot configuration
Our standard practice for YAML files is to sort keys alphabetically.
This mirrors what was done to the configurations inherited from
cisagov/skeleton-generic. This also moves the attribution comment for
the ignore directives under the `ignore` block itself to match the way
it is done fo rthe `github-actions` package ecosystem declaration.